chore(wallet-filter-sidebar): rename useHook file
